Created a branch for our manually created skeleton
We created this skeleton from a tutorial centered around making a realtime chat app.
We are modifying the javascript to fit our purposes of music synchronization.
The first thing we replaced was an ajax call that we translated into a socket emission to be handled by the server and broadcasted back to each client.